// src/app/pages/ordens/components/ordens-table/ordens.interface.ts
export interface OrdensInterface {
   idorden?: number;
   fecha?: string;
   manoObra?: number;
   subtotal?: number;
   total?: number;
   anticipo?: number;
   status?: string;
   descripcion?: string;
   vehiculoreparando_idvehiculoreparando?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}